requirements.txt
tqdm
ijson
pyyaml
pathlib
